Add new lines at end of files and semantic changes at the recommendation of Mitch
