+++ auacm/angular/package.json
@@ -1,7 +1,7 @@
 {
   "name": "auacm",
-  "version": "0.0.1",
-  "license": "MIT",
+  "version": "3.0.0",
+  "license": "Apache-2.0",
   "angular-cli": {},
   "scripts": {
     "ng": "ng",
