+++ auacm/angular/src/app/competitions/competitions.component.html
