+++ auacm/angular/src/app/edit-competition/edit-competition.component.css
