+++ auacm/angular/src/app/judge/judge.component.css
