+++ auacm/angular/src/app/models/competition.ts
