+++ auacm/angular/src/app/models/response.ts
@@ -1,0 +1,8 @@
+export class SimpleResponse {
+  constructor(success: boolean, message: string) {
+    this.success = success;
+    this.message = message;
+  }
+  success: boolean;
+  message: string;
+}
