+++ auacm/angular/src/app/view-problem/view-problem.component.ts
@@ -41,5 +41,4 @@ export class ViewProblemComponent implements OnInit {
       }
     });
   }
-
 }
