+++ auacm/angular/src/app/view-scoreboard/view-scoreboard.component.ts
@@ -177,5 +177,4 @@ export class ViewScoreboardComponent implements OnInit, OnDestroy {
       }
     }
   }
-
 }
